+++ $notes/scriptGenerator/lib/lootBags.py
@@ -1,11 +1,25 @@
-def defaultOutput(name, entries, rarity="COMMON"):
+import json
+import os
+from src import const
+
+def writeLootFile(parentId, childId, name, lootEntries, rarity="COMMON"):
+	dataFolder = os.path.join(const.data(), 'lootbags', 'recipes', parentId)
+	if not os.path.exists(dataFolder):
+		os.makedirs(dataFolder)
+	json.dump(
+		defaultOutput(name,entries(lootEntries), rarity),
+		open(os.path.join(dataFolder, childId+'.json'), 'w'),
+		indent=2
+	)
+
+def defaultOutput(name, lootEntries, rarity="COMMON"):
 	return {
 		"type": "lootbags:loot",
 		"name": name,
 		"rarity": rarity,
 		"output": {
 			"rolls": 1,
-			"entries": entries
+			"entries": lootEntries
 		}
 	}